Moved some of the login functionality from the ExploreRecipes page to the MainActivity page
